requirements.txt
discord.py==2.3.2
imapclient==2.3.1
python-dotenv==1.0.0
aiohttp==3.9.3
